chore(core): move isNxGenerator out of workspaces
